+++ packages/snapshot-legacy-importer/source/service-provider.ts
@@ -1,0 +1,16 @@
+import { Identifiers } from "@mainsail/constants";
+import { injectable } from "@mainsail/container";
+import { Providers } from "@mainsail/kernel";
+
+import { Importer } from "./importer.js";
+
+@injectable()
+export class ServiceProvider extends Providers.ServiceProvider {
+	public async register(): Promise<void> {
+		this.app.bind(Identifiers.Snapshot.Legacy.Importer).to(Importer).inSingletonScope();
+	}
+
+	public async required(): Promise<boolean> {
+		return true;
+	}
+}
